# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets/search:
    get:
      operationId: searchAssets
      tags: [assets]
      summary: Search assets
      parameters:
        - name: q
          in: query
          schema:
            type: string
          required: true
        - name: type
          in: query
          schema:
            type: string
            enum: [crypto, stock, commodity]
          required: false
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
  /assets/{symbol}:
    get:
      operationId: getAssetDetail
      tags: [assets]
      summary: Get asset detail and quote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDetail"
  /assets/{symbol}/chart:
    get:
      operationId: getAssetChart
      tags: [assets]
      summary: Get asset price chart data
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 1w, 1m, 1y, all]
          required: false
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
  /assets/list:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List available assets by type
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [crypto, stock, commodity]
          required: false
      responses:
        "200":
          description: Assets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login with passcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid passcode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users
      parameters:
        - name: kycStatus
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, not_started]
          required: false
        - name: country
          in: query
          schema:
            type: string
          required: false
        - name: search
          in: query
          schema:
            type: string
          required: false
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [balance, createdAt, name]
          required: false
        - name: limit
          in: query
          schema:
            type: integer
          required: false
        - name: offset
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{userId}:
    get:
      operationId: getAdminUserDetail
      tags: [admin]
      summary: Get detailed user information
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{userId}/kyc:
    patch:
      operationId: updateUserKycStatus
      tags: [admin]
      summary: Update user KYC status (approve/reject/flag)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycStatusUpdateRequest"
      responses:
        "200":
          description: KYC status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ChartPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required: [date, value]
    Holding:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        assetType:
          type: string
          enum: [crypto, stock, commodity]
        quantity:
          type: number
        averageCost:
          type: number
        currentPrice:
          type: number
        currentValue:
          type: number
        gainLoss:
          type: number
        gainLossPercentage:
          type: number
        logoUrl:
          type: string
          nullable: true
      required: [id, symbol, name, assetType, quantity, averageCost, currentPrice, currentValue, gainLoss, gainLossPercentage]
...
    Asset:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        assetType:
          type: string
          enum: [crypto, stock, commodity]
        currentPrice:
          type: number
        change24h:
          type: number
        changePercent24h:
          type: number
        marketCap:
          type: number
          nullable: true
        logoUrl:
          type: string
          nullable: true
      required: [symbol, name, assetType, currentPrice, change24h, changePercent24h]
    AssetDetail:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        assetType:
          type: string
          enum: [crypto, stock, commodity]
        currentPrice:
          type: number
        change24h:
          type: number
        changePercent24h:
          type: number
        high24h:
          type: number
        low24h:
          type: number
        volume24h:
          type: number
        marketCap:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        exchange:
          type: string
          nullable: true
        lastUpdated:
          type: string
        logoUrl:
          type: string
          nullable: true
      required: [symbol, name, assetType, currentPrice, change24h, changePercent24h, high24h, low24h, volume24h, lastUpdated]
    ChartData:
      type: object
      properties:
        symbol:
          type: string
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ChartPoint"
      required: [symbol, period, data]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [buy, sell, deposit, withdraw, convert]
        symbol:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        quantity:
          type: number
          nullable: true
        price:
          type: number
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
        logoUrl:
          type: string
          nullable: true
      required: [id, type, amount, status, createdAt]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        legalName:
          type: string
          nullable: true
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        profilePhotoUrl:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [not_started, pending, approved, rejected]
        investmentPreferences:
          type: array
          items:
            type: string
          nullable: true
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
      required: [id, fullName, email, phone, country, kycStatus, onboardingComplete, createdAt]
...
    UserBalance:
      type: object
      properties:
        availableCash:
          type: number
        totalPortfolioValue:
          type: number
        cryptoBalance:
          type: number
        stockBalance:
          type: number
        commodityBalance:
          type: number
        pendingDeposits:
          type: number
        pendingWithdrawals:
          type: number
      required: [availableCash, totalPortfolioValue, cryptoBalance, stockBalance, commodityBalance, pendingDeposits, pendingWithdrawals]
    AdminLoginRequest:
      type: object
      properties:
        passcode:
          type: string
      required: [passcode]
...
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [users, total, limit, offset]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        country:
          type: string
        profilePhotoUrl:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [not_started, pending, approved, rejected]
        onboardingStep:
          type: integer
        onboardingComplete:
          type: boolean
        availableBalance:
          type: number
        totalAssets:
          type: number
        holdingsCount:
          type: integer
        transactionCount:
          type: integer
        lastActive:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, fullName, email, country, kycStatus, onboardingStep, onboardingComplete, availableBalance, totalAssets, holdingsCount, transactionCount, createdAt]
    AdminUserDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        balance:
          $ref: "#/components/schemas/UserBalance"
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/Holding"
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        kycDocuments:
          type: array
          items:
            $ref: "#/components/schemas/KycDocument"
        selfieStatus:
          type: string
          enum: [not_submitted, submitted, approved, rejected]
        kycNotes:
          type: string
          nullable: true
        activityTimeline:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEvent"
      required: [user, balance, holdings, recentTransactions, kycDocuments, selfieStatus]
    KycDocument:
      type: object
      properties:
        id:
          type: integer
        documentType:
          type: string
        side:
          type: string
        fileUrl:
          type: string
        uploadedAt:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
      required: [id, documentType, side, fileUrl, uploadedAt, status]
    ActivityEvent:
      type: object
      properties:
        id:
          type: integer
        eventType:
          type: string
        description:
          type: string
        timestamp:
          type: string
      required: [id, eventType, description, timestamp]
    KycStatusUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected, flagged, pending]
        notes:
          type: string
          nullable: true
      required: [status]