# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetAssetType } from "./assetAssetType";

export interface Asset {
  symbol: string;
  name: string;
  assetType: AssetAssetType;
  currentPrice: number;
  change24h: number;
  changePercent24h: number;
  marketCap?: number | null;
  logoUrl?: string | null;
}