# lib/api-zod/src/generated/types/kycDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { KycDocumentStatus } from "./kycDocumentStatus";

export interface KycDocument {
  id: number;
  documentType: string;
  side: string;
  fileUrl: string;
  uploadedAt: string;
  status: KycDocumentStatus;
}