# lib/api-zod/src/generated/types/userBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserBalance {
  availableCash: number;
  totalPortfolioValue: number;
  cryptoBalance: number;
  stockBalance: number;
  commodityBalance: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
}